# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HRIS Okaeri API Documentation",
    "version": "1.0.0",
    "description": "API documentation for HRIS Okaeri system"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/branches": {
      "post": {
        "tags": ["Branches"],
        "summary": "Create a new branch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Kantor Pusat"
                  },
                  "code": {
                    "type": "string",
                    "example": "HO-001"
                  },
                  "address": {
                    "type": "string",
                    "example": "Jl. Main Office No. 1"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "021-12345678"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "ho@company.com"
                  },
                  "latitude": {
                    "type": "number",
                    "format": "decimal",
                    "minimum": -90,
                    "maximum": 90,
                    "example": -6.185141,
                    "description": "Branch latitude coordinate (South is negative)"
                  },
                  "longitude": {
                    "type": "number",
                    "format": "decimal",
                    "minimum": -180,
                    "maximum": 180,
                    "example": 106.831762,
                    "description": "Branch longitude coordinate (West is negative)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Branch created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "name": {
                          "type": "string",
                          "example": "Kantor Pusat"
                        },
                        "code": {
                          "type": "string",
                          "example": "HO-001"
                        },
                        "address": {
                          "type": "string",
                          "example": "Jl. Main Office No. 1"
                        },
                        "phoneNumber": {
                          "type": "string",
                          "example": "021-12345678"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "example": "ho@company.com"
                        },
                        "latitude": {
                          "type": "number",
                          "format": "decimal",
                          "example": -6.185141
                        },
                        "longitude": {
                          "type": "number",
                          "format": "decimal",
                          "example": 106.831762
                        },
                        "isActive": {
                          "type": "boolean",
                          "example": true
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-01-20T08:30:00.000Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-01-20T08:30:00.000Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/branches/{id}": {
      "put": {
        "tags": ["Branch Management"],
        "summary": "Update a branch",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Branch ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Kantor Pusat"
                  },
                  "code": {
                    "type": "string",
                    "example": "HO-001"
                  },
                  "address": {
                    "type": "string",
                    "example": "Jl. Main Office No. 1"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "021-12345678"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "ho@company.com"
                  },
                  "latitude": {
                    "type": "number",
                    "format": "decimal",
                    "minimum": -90,
                    "maximum": 90,
                    "example": -6.185141,
                    "description": "Branch latitude coordinate (South is negative)"
                  },
                  "longitude": {
                    "type": "number",
                    "format": "decimal",
                    "minimum": -180,
                    "maximum": 180,
                    "example": 106.831762,
                    "description": "Branch longitude coordinate (West is negative)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Branch updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        },
                        "address": {
                          "type": "string"
                        },
                        "phoneNumber": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Branch not found"
          }
        }
      }
    }
  }
}